+++ pimd/pim_bfd.c
@@ -28,6 +28,7 @@
 #include "zclient.h"
 
 #include "pim_instance.h"
+#include "pim_neighbor.h"
 #include "pim_cmd.h"
 #include "pim_vty.h"
 #include "pim_iface.h"
