+++ pimd/pim_oil.h
@@ -20,8 +20,9 @@
 #ifndef PIM_OIL_H
 #define PIM_OIL_H
 
+struct pim_interface;
+
 #include "pim_mroute.h"
-#include "pim_iface.h"
 
 /*
  * Where did we get this (S,G) from?
