+++ pimd/pim_zlookup.c
@@ -31,6 +31,7 @@
 
 #include "pimd.h"
 #include "pim_iface.h"
+#include "pim_neighbor.h"
 #include "pim_pim.h"
 #include "pim_str.h"
 #include "pim_oil.h"
