+++ pimd/pimd.h
@@ -136,7 +136,6 @@ extern const char *const PIM_ALL_ROUTERS;
 extern const char *const PIM_ALL_PIM_ROUTERS;
 extern const char *const PIM_ALL_IGMP_ROUTERS;
 
-extern struct pim_router *router;
 extern struct zebra_privs_t pimd_privs;
 extern struct in_addr qpim_all_pim_routers_addr;
 extern uint8_t qpim_ecmp_enable;
